load all posts here

+++ multer/multerImage.js
@@ -1,6 +1,7 @@
 
 const router = require("express").Router()
-
+const jwt = require("jsonwebtoken")
+const KEY = process.env.SECRET_KEY
 const DetectImage = require("../ImagesDetection/_detect")
 const Noti = require('../db/Notification');
 const fs = require("fs")
@@ -13,6 +14,7 @@ const Post = require("../db/UserData");
 const Comments = require("../db/Comments");
 const TextPost = require("../db/TextPost");
 const Model = require("../public/Nsfw_Model/min_nsfwjs/model.json")
+
 
 let Pusher = require('pusher');
 let pusher = new Pusher({
@@ -36,8 +38,8 @@ let pusher = new Pusher({
 // pusher.trigger("userDetails", "message1", {
 // });
 
-pusher.trigger("updateComment", "updateCommentMessage", {
-});
+// pusher.trigger("updateComment", "updateCommentMessage", {
+// });
 
 const url = "https://cdni.pornpics.com/460/7/433/79976277/79976277_257_72b0.jpg"
 
@@ -67,8 +69,12 @@ const upload = multer({ storage: storage }).single("file")
 router.post("/user/blob/image/9fHtqOJumtxOzmTfLMFT/ETXsG3rHrnx2irUZmefU/njVzxxrEx84ZrUiERB0t/fxXRdJLMKIkzxucTbovy/sO9rLr3E0EuDpjYcawQD/:id", async (req, res) => {
     try {
         // const { _id } = req.user
-        const _id = req.params.id
+        const token = req.params.id
         const { data, url } = req.body
+        console.log({ token })
+
+        const { _id } = await jwt.verify(token, KEY)
+        console.log(_id)
         //filter the local image url
         const filterUrl = url.split("blob")[1].slice(1)
 
@@ -129,7 +135,7 @@ router.post("/user/blob/image/9fHtqOJumtxOzmTfLMFT/ETXsG3rHrnx2irUZmefU/njVzxxrE
 
     }
     catch (err) {
-        return res.status(499).json({ message: "not uploaded please try again" })
+        return res.status(499).json({ message: "not uploaded please try again" + err })
     }
 })
 
@@ -141,10 +147,12 @@ router.get("/profile/image/e9thhvkKqJpnTlYo1sQl/QVbghZqhoSr2Rt5qvNYJ/iKj3RoJojFW
     try {
 
         // const { _id } = req.user
-        const _id = req.params.id
+        const token = req.params.id
 
+        // console.log({ token })
 
-
+        //verify token for current user
+        const { _id } = await jwt.verify(token, KEY)
 
         // UserBlob/
         const result = await cloudinary.search.expression(
@@ -238,13 +246,18 @@ router.delete("/delete/assest/:id", async (req, res) => {
 
 
 
-router.post("/user/i/b/y9y5y0q3eztm3ibcd8z0/bum6ozd9m1sw4w9fbxea/amqvdkbe49sn4u3cvsvt/e5ce6ba3miamapdl7wyv", async (req, res) => {
+
+//save usr information
+
+router.post("/user/i/b/y9y5y0q3eztm3ibcd8z0/bum6ozd9m1sw4w9fbxea/amqvdkbe49sn4u3cvsvt/e5ce6ba3miamapdl7wyv/:id", async (req, res) => {
     try {
 
-        const { _id } = req.user
+        // const { _id } = req.user
+        const token = req.params.id
         const { username, fname, lname, gender, address, city, country, postalCode, college, stream, degree, position, aboutMe } = req.body
 
 
+        const _id = await jwt.verify(token, KEY)
 
         if (!username || !fname || !lname || !gender || !address || !city || !country || !postalCode || !aboutMe || !college || !stream || !degree || !position) {
             res.status(401).json({ message: "Please fill all the fields" })
@@ -324,7 +337,14 @@ router.post("/user/i/b/y9y5y0q3eztm3ibcd8z0/bum6ozd9m1sw4w9fbxea/amqvdkbe49sn4u3
 router.get("/user/083525p7ljhwmxifts31/l66cbrsuytmj1wujuauz/nqoye5ozdqj89b4s4qoq/ua1iztaxjo4bbmzvd391/3mzqeygnoszlknp90h51/t28uf00khscofxgjwj20/:id", async (req, res) => {
     try {
 
-        const _id = req.params.id
+
+
+        // const _id = req.params.id
+
+        const token = req.params.id
+
+        //verify token for current user
+        const { _id } = await jwt.verify(token, KEY)
         const userInformationLoadFromServer = await Post.findOne({ googleId: _id })
 
 
@@ -350,7 +370,9 @@ router.post("/user/blob/image/bg/S6MjFqeb8HdJRGjkUs9W/QUCzIb1mKtMevddN24yB/YWYht
 
         // const { _id } = req.user
         const { data } = req.body
-        const id = req.params.id
+        const token = req.params.id
+        const { _id } = await jwt.verify(token, KEY)
+        const id = _id
         const uploadResponse = await cloudinary.uploader.upload(data, {
             folder: `${id}/background`,
             public_id: id,
@@ -410,10 +432,16 @@ router.get("/bg/image/mwQgga2z5KfChXjuF1s0/r6dg0LqqWmCG4W5UQOTa/ftFhzft7YNwT6jb9
 
 
 
-        const id = req.params.id
-        // console.log(id)
+        // const id = req.params.id
+        // console.log({id})
+        const token = req.params.id
+
+        //verify token for current user
+        const { _id } = await jwt.verify(token, KEY)
+
+
         const result = await cloudinary.search.expression(
-            "folder:" + `${id}/background`,
+            "folder:" + `${_id}/background`,
         )
             .sort_by('created_at', 'desc')
             // .max_results(20)
@@ -474,6 +502,7 @@ router.get("/root/load/all/comments/:commentId/:userId", async (req, res) => {
         // const { _id } = req.user
         // console.log(req.params)
         const { commentId, userId } = req.params
+        console.log({ commentId, userId })
         const AllUsersComments = await Comments.find({
             $and: [{ post_id: commentId }, {
                 userId
@@ -588,6 +617,7 @@ router.post("/users/post/:id", async (req, res) => {
         let array = []
         const postId = req.params.id
         const { text, image, privacy, post_id, time } = req.body
+        console.log({ image })
         const data = image
         const userId = postId.split("-")[0]
         if (image || text) {
@@ -766,6 +796,7 @@ router.get("/users/public/posts/:id", async (req, res) => {
     try {
         let array = []
         // const { id } = req.query
+        console.log("isd", { _id: req.params.id })
         const result = await cloudinary.search.expression(
             "folder:" + `${req.params.id}/post`,
         ).execute()
@@ -885,17 +916,23 @@ router.delete("/delete/user/post/:id", async (req, res) => {
 
 
 
+
+
+
+
 //==================================save user post into the mongodb  by local url=========================
 
 router.post("/local/url/:id", async (req, res) => {
     try {
         // console.log(res.body)
-        const id = req.params.id
-        const { text, image, privacy, post_id, time, fileType, likes_count, liked } = req.body
-        // console.log("user post")
-        // console.log(req.body)
-        // console.log("file type")
-        // console.log(fileType)
+        const token = req.params.id
+        const { _id } = await jwt.verify(token, KEY)
+        const id = _id
+        const { text, image, name, privacy, post_id, time, fileType, likes_count, liked, userProfileImageUrl } = req.body
+        console.log("user post")
+        console.log(req.body)
+        console.log("file type")
+        console.log(fileType)
 
 
         //filter the local image url
@@ -905,15 +942,18 @@ router.post("/local/url/:id", async (req, res) => {
 
 
         const SaveUserPosts = await new TextPost({
-            username: "SAnju",
+            username: name,
             text: text,
             image: image,
             fileType: fileType,
             privacy: privacy,
             post_id: post_id,
             userId: id,
+            profileImage: userProfileImageUrl,
             likes_count: likes_count,
             liked: liked,
+            post_url: "http://localhost:5000/blob/user/post/" + post_id,
+
             createdAt: time,
 
         })
@@ -927,7 +967,8 @@ router.post("/local/url/:id", async (req, res) => {
                 return res.status(500).json({ message: "Not Post" + err })
             }
             else {
-                const GetAllUserPost = await TextPost.find({})
+                //send all user post by Id
+                const GetAllUserPost = await TextPost.find({ $or: [{ userId: id }, { privacy: "public" }] })
                 await pusher.trigger("AddPost", "AddPostMessage", {
                     GetAllUserPost: GetAllUserPost.reverse()
                 }, req.body.socketId);
@@ -947,6 +988,21 @@ router.post("/local/url/:id", async (req, res) => {
 
 
 
+//Each post url for share the post or download the post
+router.get("/user/post/:post_id", async (req, res) => {
+    try {
+        const { post_id } = req.params
+
+        const post = await TextPost.findOne({ post_id: post_id })
+        return res.status(200).json({ post_url: post })
+
+
+    } catch (err) {
+        return res.status(500).json({ message: "Something error occured" })
+
+    }
+})
+
 
 
 
@@ -955,8 +1011,14 @@ router.post("/local/url/:id", async (req, res) => {
 
 router.get("/load/all/post/:id", async (req, res) => {
     try {
-        const id = req.params.id
-        const GetAllUserPost = await TextPost.find({})
+        const token = req.params.id
+        const { _id } = await jwt.verify(token, KEY)
+
+        //get all post by userId 
+        const GetAllUserPost = await TextPost.find({ $or: [{ userId: _id }, { privacy: "public" }] })
+        const GetAllUserPost1 = GetAllUserPost.filter((item) => {
+            return
+        })
         return res.status(200).json({ message: "successfull load", data: GetAllUserPost.reverse() })
     } catch (error) {
 
@@ -970,7 +1032,14 @@ router.get("/load/all/post/:id", async (req, res) => {
 router.delete("/delete/user/post/local/:id", async (req, res) => {
     try {
         const id = req.params.id
-        const deletePost = await TextPost.findOneAndDelete({ post_id: id })
+        console.log({ id })
+        const { userId } = req.body
+
+        console.log("user deleletd this")
+        console.log({ userId })
+        console.log({ id })
+        const deletePost = await TextPost.findOneAndDelete({ $and: [{ post_id: id }, { userId: userId }] })
+        await Noti.findOneAndDelete({ $and: [{ post_id: id }, { userId: userId }] })
 
 
 
@@ -986,10 +1055,16 @@ router.delete("/delete/user/post/local/:id", async (req, res) => {
             //getallpost after delete
             const GetAllPostsAfterDelete = await TextPost.find({})
             console.log({ GetAllPostsAfterDelete })
-            await Noti.findOneAndDelete({ post_id: id })
+
+            const allNotiAfterDelete = await Noti.find({})
+
             pusher.trigger("DeletePost", "PostDeleted", { GetAllPostsAfterDelete: GetAllPostsAfterDelete.reverse() }, req.body.socketId)
+            pusher.trigger("DeleteNotiByPost", "DeleteNotiMessage", { allNotiAfterDelete }, req.body.socketId)
 
             return res.status(200).json({ message: "Post deleted successfully", data: GetAllPostsAfterDelete.reverse() })
+        }
+        else {
+            return res.status(200).json({ message: "you can not delete this." })
         }
 
     } catch (error) {
@@ -1003,9 +1078,15 @@ router.delete("/delete/user/post/local/:id", async (req, res) => {
 
 router.get("/all/comment/user/:id", async (req, res) => {
     try {
-        const id = req.params.id
+        // const id = req.params.id
+        // console.log({ id })
+        const token = req.params.id
+
+        //verify token for current user
+        const { _id } = await jwt.verify(token, KEY)
+
         const { post_id } = req.body
-        const GetAllComment = await Comments.find({ userId: id })
+        const GetAllComment = await Comments.find({ userId: _id })
         await pusher.trigger("updateComment", "updateCommentMessage", { GetAllComment: GetAllComment.reverse() }, req.body.socketId)
         return res.status(200).json({ message: "successfull load", data: GetAllComment.reverse() })
 
@@ -1022,24 +1103,19 @@ router.get("/all/comment/user/:id", async (req, res) => {
 
 router.put("/user/like/:id/:post_id", async (req, res) => {
     try {
+
         const { id, post_id } = req.params
+        console.log({ id })
+        console.log(post_id)
         const FindPostById = await TextPost.findOne({ post_id })
+
+
 
         if (!FindPostById.liked.includes(id)) {
 
-            const SaveNoti = await Noti({
-                name: fname + " " + lname,
-                url: url,
-                post_id: post_id,
-            })
-            SaveNoti.save((err) => {
-                if (err) {
-                    console.log("noti not saved", err)
-                }
-                else {
-                    console.log("noti saved")
-                }
-            })
+
+
+
             TextPost.findOneAndUpdate({ post_id }, {
                 $push: {
                     liked: id
@@ -1051,8 +1127,7 @@ router.put("/user/like/:id/:post_id", async (req, res) => {
                 // UserBlob/
 
 
-                //fetch userDetails
-                const { fname, lname } = await Post.findOne({ googleId: id })
+
 
                 //fetch the user image which like post
                 const result = await cloudinary.search.expression(
@@ -1061,7 +1136,27 @@ router.put("/user/like/:id/:post_id", async (req, res) => {
 
                 if (result.resources.length > 0) {
                     const url = result.resources[0].url
-                    pusher.trigger("LikePost", "LikePostMessage", { url, name: fname + " " + lname }, req.body.socketId)
+                    //fetch userDetails
+
+                    const { fname, lname } = await Post.findOne({ googleId: id })
+                    const SaveNoti = await Noti({
+                        name: fname + " " + lname,
+                        url: url,
+                        post_id: post_id,
+                        userId: id
+                    })
+
+                    SaveNoti.save(async (err) => {
+                        if (err) {
+                            console.log("noti not saved", err)
+                        }
+                        else {
+                            console.log("noti saved")
+                            const allNoti = await Noti.find({})
+                            pusher.trigger("LikePost", "LikePostMessage", { url, name: fname + " " + lname, allNoti }, req.body.socketId)
+                        }
+                    })
+
 
 
                 }
@@ -1099,12 +1194,14 @@ router.put("/user/like/:id/:post_id", async (req, res) => {
 
                 if (result.resources.length > 0) {
                     const url = result.resources[0].url
-                    pusher.trigger("LikePost", "LikePostMessage", { url, name: fname + " " + lname }, req.body.socketId)
+                    const allNoti = await Noti.find({})
+                    pusher.trigger("LikePost", "LikePostMessage", { url, name: fname + " " + lname, allNoti }, req.body.socketId)
 
 
                 }
                 else {
-                    pusher.trigger("LikePost", "LikePostMessage", { url, name: fname + "" + lname }, req.body.socketId)
+                    const allNoti = await Noti.find({})
+                    pusher.trigger("LikePost", "LikePostMessage", { url, name: fname + "" + lname, allNoti }, req.body.socketId)
 
 
 
@@ -1117,50 +1214,6 @@ router.put("/user/like/:id/:post_id", async (req, res) => {
 
 
 
-
-
-
-
-
-
-        // console.log("user like with deffrenet posts")
-        // // console.log(req.params)
-        // // console.log(req.body)
-        // const { action, userId } = req.body
-        // const counter = action === true ? 1 : -1
-        // // console.log("action gadi")
-        // // console.log(counter)
-
-        // TextPost.findOneAndUpdate({ post_id: post_id }, { $inc: { likes_count: counter }, $push: { liked: req.body } }, { new: true }, async (err, result) => {
-        //     if (err) {
-        //         // console.log(err)
-        //         return res.status(500).json({ message: "not liked" })
-        //     }
-        //     else {
-        //         //filter all poost
-        //         const getAllpost = await TextPost.find({})
-
-
-        //         // console.log(result)
-        //         // find user details which liked post
-        //         const userDetailsFind = await Post.findOne({ googleId: userId })
-        //         const UserProfileImage = await cloudinary.search.expression(
-        //             "folder:" + userId + "/profileImage")
-        //             .sort_by('created_at', 'desc')
-        //             .execute()
-        //         pusher.trigger('channel', 'message', {
-        //             action: action, post_id: post_id, likedUser: userId, result: getAllpost.reverse()
-        //         }, req.body.socketId);
-        //         pusher.trigger('userDetails', 'message1', {
-        //             userDetailsFind,
-        //             UserProfileImage
-
-        //         }, req.body.socketId);
-        //         return res.status(200).json({ message: "Successfully like", data: result })
-        //     }
-
-        // })
-
     } catch (err) {
         return res.status(500).json({ message: "Something error occured" + err })
 
@@ -1172,6 +1225,8 @@ router.put("/user/like/:id/:post_id", async (req, res) => {
 router.get("/load/userliked/details/:id", async (req, res) => {
     try {
         const userId = req.params.id
+        console.log({ userId })
+        console.log("loked")
 
         // find user details which liked post
         const userDetailsFind = await Post.findOne({ googleId: userId })
@@ -1221,10 +1276,12 @@ router.get("/search/", async (req, res) => {
 
 
 //load the all notification
-router.get("/load/notification/:id", async (req, res) => {
+router.get("/load/all/notification/:id", async (req, res) => {
     try {
-        const { id } = req.params
-        const result = await Noti.find({})
+        const token = req.params.id
+        const { _id } = await jwt.verify(token, KEY)
+        const result = await Noti.find({ userId: _id })
+        console.log({ result })
         return res.status(200).json({ message: "successfull", data: result })
 
 
@@ -1240,12 +1297,13 @@ router.put("/visibility/user/post/local/:post_id", async (req, res) => {
     try {
         const { post_id } = req.params
         const { visibility } = req.body
-      TextPost.findOneAndUpdate({ post_id }, { privacy: visibility }, { new: true }, async (err, result) => {
+        TextPost.findOneAndUpdate({ post_id }, { privacy: visibility }, { new: true }, async (err, result) => {
             if (err) {
                 return res.status(500).json({ message: "somethinng error occured" + err })
             }
             else {
-                return res.status(200).json({ message: "successfull", data: result })
+                const allPost = await TextPost.find({})
+                return res.status(200).json({ message: "successfull", data: allPost })
             }
         })
 
@@ -1259,4 +1317,25 @@ router.put("/visibility/user/post/local/:post_id", async (req, res) => {
 
 })
 
+
+
+//update the privay of any post
+router.put("/api/setPrivacy/:post_id", async (req, res) => {
+    try {
+        const { post_id } = req.params
+        const { privacy } = req.body
+        TextPost.findOneAndUpdate({ post_id }, { privacy }, { new: true }, async (err, result) => {
+            if (err) {
+                return res.status(500).json({ message: "somethinng error occured" + err })
+            }
+            else {
+                const allPost = await TextPost.find({})
+                return res.status(200).json({ message: "successfull", data: allPost })
+            }
+        })
+    }
+    catch (err) {
+        return res.status(500).json({ message: "somethinng error occured" + err })
+    }
+})
 module.exports = router;
